+++ matching_game/backend/src/handler/game.go
@@ -1,50 +1,27 @@
 package handler
 
 import (
-	"encoding/json"
 	"fmt"
-	"log"
 	"net/http"
+	"strings"
 
 	"github.com/SuperTikuwa/matching_game/redisoperation"
-	"github.com/gorilla/mux"
 )
 
-func GameGetHandler(w http.ResponseWriter, r *http.Request) {
-	g := redisoperation.GetGames()
-
-	games := struct {
-		Games []string `json:"games"`
-	}{
-		g,
-	}
-
-	json, err := json.Marshal(games)
-	if err != nil {
-		log.Println(err)
-	}
-
-	fmt.Fprint(w, string(json))
-}
-
 func GamePostHandler(w http.ResponseWriter, r *http.Request) {
-	gameID := redisoperation.GenerateGameID()
+	gameID := redisoperation.FindEmptyGame()
+	if gameID == "" {
+		gameID = redisoperation.GenerateGameID()
+		redisoperation.CreateGame(gameID, "12345")
 
-	j := struct {
-		ID string `json:"gameID"`
-	}{
-		gameID,
+		w.WriteHeader(http.StatusCreated)
+		fmt.Fprint(w, `{"gameID":"`+gameID+`"}`)
+		return
 	}
 
-	res, err := json.Marshal(j)
-	if err != nil {
-		log.Fatal(err)
-	}
+	gameID = strings.Split(gameID, ":")[1]
 
-	fmt.Fprintln(w, string(res))
-}
-
-func GamePutHandler(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	redisoperation.JoinGame(vars["game_id"], r.URL.Query().Get("player_id"))
+	redisoperation.JoinGame(gameID, "54321")
+	w.WriteHeader(http.StatusOK)
+	fmt.Fprint(w, `{"gameID":"`+gameID+`"}`)
 }
