+++ matching_game/backend/src/main.go
@@ -14,10 +14,8 @@ func init() {
 func main() {
 	r := mux.NewRouter()
 
-	r.HandleFunc("/words", handler.WordsGetHandler).Methods("GET")
+	r.HandleFunc("/words", handler.WordsGetHandler).Methods("GET").Queries("mode", "{mode:[0-2]}", "gameID", "{gameID}")
 
-	r.HandleFunc("/games", handler.GameGetHandler).Methods("GET")
 	r.HandleFunc("/games", handler.GamePostHandler).Methods("POST")
-	r.HandleFunc("/games/{game_id}", handler.GamePutHandler).Methods("PUT")
 	http.ListenAndServe(":80", r)
 }
